Crate deskulpt_common

This crate implements shared utilities for the Deskulpt application.

âš ï¸ This crate is meant to be used internally by the Deskulpt application and is not designed to support plugin authors or other users directly. Private items are documented for reference of Deskulpt developers.

Modules

bindings

Common utilities for declaring Deskulpt bindings.

event

Common utilities for Deskulpt events.

init

Common utilities for initializing internal Tauri plugins for Deskulpt.

outcome

ser_error ğŸ”’

window

Common utilities for Deskulpt windows.

Macros

ser_bail

anyhow::bail! that returns early with a SerError.

Structs

SerError

Serializable wrapper around anyhow::Error.

Type Aliases

SerResult

Result type with SerError.