deskulpt_common

Module bindings

Common utilities for declaring Deskulpt bindings.

Macros

build_bindings

Create a function that builds Bindings for this crate.

collect_commands

Used in BindingsBuilder::commands.

Structs

Bindings

A collection of types, events, and commands to be exposed to the frontend.

BindingsBuilder

Builder for a Bindings instance.