deskulpt_common::init

Macro init_builder

macro_rules! init_builder {
    () => { ... };
}

Initialize a tauri::plugin::Builder.

The builder has its name automatically set to the crate name, and its invoke handler set to the commands specified in the build script. It can be further customized before calling the .build() method. One must configure the build script correctly with deskulpt-build.